fix(backwards-compatible): adjustments so kong-build-tools is backwards compatible and a CI check to verify as such (#560)
* fix(backwards-compatible): adjustments so we're backwards compatible with 2.8.0

* fix(tests): keep the tests just make them conditional

* chore(ci): make our build process less verbose [reverts 92e4d4b]

* fix(syntax): derpy syntax fix

+++ build-kong.sh
@@ -52,12 +52,14 @@ pushd /kong
 
   # /usr/local/kong/include is usually created by other C libraries, like openssl
   # call mkdir here to make sure it's created
-  mkdir -p /tmp/build/usr/local/kong/include
-  cp -r kong/include/* /tmp/build/usr/local/kong/include/
+  if [ -e "kong/include" ]; then
+    mkdir -p /tmp/build/usr/local/kong/include
+    cp -r kong/include/* /tmp/build/usr/local/kong/include/
+  fi
 
   # circular dependency of CI: remove after https://github.com/Kong/kong-distributions/pull/791 is merged
   if [ -e "kong/pluginsocket.proto" ]; then
-        cp kong/pluginsocket.proto /tmp/build/usr/local/kong/include/kong
+    cp kong/pluginsocket.proto /tmp/build/usr/local/kong/include/kong
   fi
 
   curl -fsSLo /tmp/protoc.zip https://github.com/protocolbuffers/protobuf/releases/download/v3.19.0/protoc-3.19.0-linux-x86_64.zip
